Move SessionsImpl to sdk package for consistency; Move saveSession extension to call site
