+++ sdk/src/main/java/com/stytch/sessions/SessionStorage.kt
@@ -1,9 +1,6 @@
 package com.stytch.sessions
 
 import com.stytch.sdk.StytchClient
-import com.stytch.sdk.StytchExceptions
-import com.stytch.sdk.StytchResult
-import com.stytch.sdk.network.responseData.IAuthData
 import com.stytch.sdk.network.responseData.SessionData
 
 private const val PREFERENCES_NAME_SESSION_JWT = "session_jwt"
@@ -60,18 +57,4 @@ internal class SessionStorage {
             session = null
         }
     }
-}
-
-//    save session data
-internal fun <T : IAuthData> StytchResult<T>.saveSession(): StytchResult<T> {
-    if (this is StytchResult.Success) {
-        value.apply {
-            try {
-                StytchClient.sessionStorage.updateSession(sessionToken, sessionJwt, session)
-            } catch (ex: Exception) {
-                return StytchResult.Error(StytchExceptions.Critical(ex))
-            }
-        }
-    }
-    return this
 }
